+++ CHANGES.md
@@ -6,6 +6,7 @@ This file marks the changes in each version.
 #### Additions
 - Add in `setRepositoryDescription` method to set the short/full description for your own repositories.
 - Add in `webhooks` method to get the webhooks for a repository you own.
+- Add in `createWebhook` method to create a webhook for a repository you own.
 
 ## 0.1
 ### 0.1.0